Overwrite `ChallengeScreen.kt` with `NewChallengeScreen.kt`
